python-ai/requirements.txt
fastapi
uvicorn
opencv-python
rembg
Pillow
